refactor useEntriesDetailed
